Print a warning if kimagecache.h is used without KF5::CoreAddons
If an application links against KF5::GuiAddons but not KF5::CoreAddons,
they cannot use KImageCache.  So we use defines (set by the exported
targets) to check whether the current target has been linked against
KF5::CoreAddons, and print a warning if not (to explain the expected
"kshareddatacache.h not found" error).

REVIEW: 114701