Make KFontUtils::adaptFontSize be a bit more exact
Not all fonts are linear in drawn size against their point size, the old code assumes so and simply guesses the next font size taking into account the pointSize we drawn, the geometry we got and the geometry we want. The new code is a bit slower but will make sure that we get better pointSize results in non linear fonts.

REVIEW: 119777