Fix potential crash in wayland inhibitor teardown
The manager is bound by the ClientExtension when the global is
announced. If we're not bound there is nothing to destroy.

This could happen if run against a wayland compositor without keyboard
inhibition support or managed to close before the registry is announced.

We do not need a matching guard in ShortcutsInhibitor as that is bound
on creation.

BUG: 429267