Add a fallback handler for the geo: URI scheme
Technically this is a tiny helper program that parses geo: URIs and
translates them to e.g. OSM website URLs. The URL patterns are provided
as arguments, so we can support arbitrary many services that way. Support
for wheelmap.org is included to demonstrate this, proprietary services
like Google Maps can be added in the same way if desired.

The goal of this is to allow applications to rely on geo: URIs to be
handled unconditionally, independent of whether Marble or another map
application is installed. This would allow a number of applications to
drop duplicated platform-specific custom code, such as KAddressBook,
KMail, Kalendar and Itinerary.

The second piece for this is the ability to select the preferred geo: URI
handler in systemsettings "Default Applications" module, which has been
integrated meanwhile.