waylandclipboard: Be a bit smarter about when to use the QGuiApplication::clipboard
The newly introduced keyboardFocusWatcher suffers from a similar problem on keyboard
leave as the one it fixed on keyboard enter.
When  Qt hides a window,  the focusWatcher and itself think that the
app has focus because keyboard_leave did not arrive yet but setting the
normal clipboard will still fail. This is an issue in klipper as it shows this
pattern of window hide and afterwards setting the clipboard.
To workaround this we can check if the focusWindow is visible even though
we still think we should have focus.
BUG:462979
FIXED-IN:5.102