clipboard: Add roundtrip when setting clipboard
To avoid deadlocking when copying data, klipper will use the regular
clipboard whilst we have keyboard focus. This is a bit of a hack, but
without better alternatives.

If data is set this way, the check when we receive a new clipboard to
see if we are the owner does not work. Checking
QClipboard::ownsClipboard does not work as we still want to track
changes from other parts of the smae process.

As an alternative, when using this specific path we make it synchronous
so that the change signal can be ignored.

BUG: 466414
FIXED-IN: 6.3