Simplify clipboard
On wayland blah blah...

Having two clipboards in use is incredibly messy, especially as it was
trivial to deadlock each other due to the syncronous mime reads with the
QMimeData API.

As a hack to avoid deadlocking whenever had keyboard focus we copied the
mimedata to the main clipboard whenever we had focus.

Following XYZ in kwin, we can build on that idea. By always using the
normal clipboard for writing, we only need to implement half the
protocol for listening to clipboard events. The code then becomes
trivial and easy to follow.