recorder: Add KInputSequence
This class has the ability to hold keyboard and mouse events alike. This
will enable us to tie them together and pass them through APIs like a
future recorder that can record these events in parallel.

This is to enable future workarounds in the limitations of
KKeySequenceRecorder which only handles keystrokes. This is important
for drawing tablet users who want to bind mouse buttons to their pad
buttons, and not just keyboard events.

CCBUG: 469232