waylandclipboard: Dont explicitly clear when transfering sources
KWin recently changed checks whether clients are allowed to change
clipboard to be more in line with other compositors, it now uses
the serial in the set_selection request to allow or decline it.
By using the data_control request we effectively invalidate our
serial since it refers to the state before the clipboard was
explicitly set. To fix this we can simply delete the  data_control
source and not explicitely set the selection.