Introduce KColorManipulation
KColorUtils is built on an obscure color space, and has
undesirable results stemming from inadequacies of the
selected color space. The operations are also anchored
on absolute points rather than relative to the input
colors, which results in the strength of effects
varying dramatically based on the distance to the
set points.

KColorManipulation is based on OKLAB, which is a more
widely understood perceptual colorspace, as well as
avoiding the use of fixed setpoints where possible.
This code is battle-tested from use in plasma-workspace
where it powers accent colour and tinting functionality.