WaylandClipboard: make sure format list doesn't have duplicate items
For GTK applications running on Xwayland it's expected for them to send
UTF8_STRING and text/plain;charset=utf-8 in one offer at the same time
due to some compatiblity reasons, but the compositor will convert
UTF8_STRING to text/plain;charset=utf-8, which causes duplications.

Given that duplicate formats are not useful in QMimeData, check a mime
type already exists before adding it to the format list.