systemclipboard: Cache mimetypes after retrieval
For a given data offer there is no need to fetch the same data multiple
types for the same mimetype. At best it's wasteful, at worse some
applications quit when their mimedata is fetched such as wl-paste or
spectacle and the second time will fail.

This brings the clipboard implementation in line with upstream Qt. This
is important as from Qt6.9 this backend can be dropped.

BUG: 497735