Add KKeySequenceRecorder patterns
The KKeySequenceRecorder supports the following shortcut formats:

- key
- modifier + key
- modifier + optional key

If it's desired to record only modifiers, for example in order to
specify what keyboard modifiers must be pressed when performing a
certain operation, the last format should be selected, i.e.
modifier + optional key. But then the problem is that the key
sequence recorder is not going to filter out normal keys, for example
inputs like "Meta+A" will be accepted, which is going to look odd.

In order to support the modifier only shortcut case, a new mode is
needed. It can be implemented by introducing a new property, such as
`modifierOnly`, but that's going to make the key sequence recorder
api more confusing because there are going to be three properties
to control the shortcut format: modifierOnly, modifierlessAllowed, and
modifierOnlyAllowed.

This change introduces the patterns property, which is a list of
accepted shortcut formats. It provides a unified way how to specify the
shortcut format.

* if the user wants to record only modifiers, use `Modifier`
* if the user wants to record a modifier only shortcut, or a normal
  shortcut, use `Modifier | ModifierAndKey`
* if the user wants to record a key, use `Key`