systemclipboard: use QFile to write data to file descriptor
Someone in the bugreport says CopyQ is better than Plasma's
clipboard manager. I checked the difference and CopyQ uses
QFile to write data to a file descriptor. QFile will check if
all bytes are written, and when not all bytes are written and
there is no error, QFile will continue writing to the file descriptor.

CCBUG: 498341